refactor(adapter): Refactor adapter utils
